+++ lib/screens/home/home_screen.dart
@@ -3,7 +3,7 @@ import 'package:gap/gap.dart';
 import 'package:jiffy/jiffy.dart';
 import 'package:live_frontend/models/clover_mission_model.dart';
 import 'package:live_frontend/models/mission_models.dart';
-import 'package:live_frontend/screens/home/widgets/clover_mission/clover_mission_list.dart';
+import 'package:live_frontend/screens/home/widgets/clover_mission_list.dart';
 import 'package:live_frontend/screens/home/widgets/clover_mission/new_clover_mission_modal.dart';
 import 'package:live_frontend/screens/home/widgets/home_profile.dart';
 import 'package:live_frontend/theme/app_colors.dart';
